move test code to Test module.
